Use proper types for longs.
Refs #85.

+++ source/demyo-service/src/main/java/org/demyo/service/impl/FilePondService.java
@@ -28,9 +28,9 @@ import org.demyo.utils.io.DIOUtils;
 @Service
 public class FilePondService implements IFilePondService {
 
-	private static final int AUTOCLEAN_DELAY = 1000 * 60;
-	private static final int AUTOCLEAN_PERIOD = 1000 * 60 * 60 * 24;
-	private static final long AUTOCLEAN_MIN_AGE = 1000 * 60 * 60;
+	private static final long AUTOCLEAN_DELAY = 1000L * 60;
+	private static final long AUTOCLEAN_PERIOD = 1000L * 60 * 60 * 24;
+	private static final long AUTOCLEAN_MIN_AGE = 1000L * 60 * 60;
 	private static final Logger LOGGER = LoggerFactory.getLogger(FilePondService.class);
 	private static final String UPLOAD_DIRECTORY_NAME = "filepond";
